Parameterized SortingBar and added to timeline
