+++ ui/src/components/Notification/NotificationListFilter.jsx
@@ -6,8 +6,9 @@ import { connect } from 'react-redux';
 import { Alignment, Button, Intent, MenuItem } from '@blueprintjs/core';
 
 import { Count, SelectWrapper } from 'components/common';
+import SortingBar from 'components/SortingBar/SortingBar';
+import SortingBarSelect from 'components/SortingBar/SortingBarSelect';
 
-import 'src/components/common/SortingBar.scss';
 
 const messages = defineMessages({
   filter_all: {
@@ -21,15 +22,6 @@ class NotificationListFilter extends Component {
     super(props);
     this.onSelect = this.onSelect.bind(this);
   }
-
-  renderOption = (option, { handleClick }) => (
-    <MenuItem
-      key={option.field || 'all'}
-      onClick={handleClick}
-      text={option.label}
-      label={<Count count={option.count} />}
-    />
-  )
 
   onSelect(selected) {
     const { query, updateQuery } = this.props;
@@ -49,41 +41,15 @@ class NotificationListFilter extends Component {
     const activeItem = typeOptions.find(item => item.active) || defaultOption;
 
     return (
-      <div className="SortingBar">
-        <span className="SortingBar__label">
-          <FormattedMessage
-            id="notifications.type_filter"
-            defaultMessage="Show:"
-          />
-        </span>
-        <div className="SortingBar__control">
-          <SelectWrapper
-            itemRenderer={this.renderOption}
+      <SortingBar
+        filterButton={(
+          <SortingBarSelect
             items={typeOptions}
-            onItemSelect={this.onSelect}
+            onSelect={this.onSelect}
             activeItem={activeItem}
-            popoverProps={{
-              minimal: true,
-              fill: false,
-              className: 'SortingBar__item__popover',
-            }}
-            inputProps={{
-              fill: false,
-            }}
-            filterable={false}
-            resetOnClose
-            resetOnSelect
-          >
-            <Button
-              text={activeItem?.label}
-              alignText={Alignment.LEFT}
-              minimal
-              intent={Intent.PRIMARY}
-              rightIcon="caret-down"
-            />
-          </SelectWrapper>
-        </div>
-      </div>
+          />
+        )}
+      />
     );
   }
 }
