+++ ui/src/components/SortingBar/SortingBar.scss
