+++ ui/src/screens/TimelineScreen/TimelineScreen.jsx
@@ -105,6 +105,7 @@ const mapStateToProps = (state, ownProps) => {
   const entitiesQuery = entitySetEntitiesQuery(location, entitySetId, null, 1000)
     .add('facet', 'dates')
     .add('facet_interval:dates', 'year')
+    .defaultSortBy('properties.date', 'asc')
 
   return {
     entitySetId,
